BACKPORT: arm64: introduce mm context flag to keep 32 bit task information
[ Upstream commit 06beb72fbe23e348cc423dd9310c6fc02cd7c7da ]

We need to decide in some cases like uprobe instruction analysis that
whether the current mm context belongs to a 32 bit task or 64 bit.

This patch has introduced an unsigned flag variable in mm_context_t.
Currently, we set and clear TIF_32BIT depending on the condition that
whether an elf binary load sets personality for 32 bit or 64 bit
respectively.

[ Mathias Gluszczynski: Preserve the downstream CONFIG_VDSO32 selection of
_SET_AUX_ENT_VDSO while extending COMPAT_SET_PERSONALITY(). ]

Signed-off-by: Pratyush Anand <panand@redhat.com>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>