net: make ndo_get_stats64 a void function
The network device operation for reading statistics is only called
in one place, and it ignores the return value. Having a structure
return value is potentially confusing because some future driver could
incorrectly assume that the return value was used.

Fix all drivers with ndo_get_stats64 to have a void function.

Signed-off-by: Stephen Hemminger <sthemmin@microsoft.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

Conflicts:
	drivers/net/bonding/bond_main.c
	drivers/net/ethernet/alacritech/slicoss.c
	drivers/net/ethernet/broadcom/tg3.c
	drivers/net/ethernet/freescale/dpaa/dpaa_eth.c
	drivers/net/ethernet/mellanox/mlx5/core/en_main.c
	drivers/net/ethernet/mellanox/mlx5/core/en_rep.c
	drivers/net/ethernet/sfc/falcon/efx.c
	include/linux/netdevice.h
	net/xfrm/xfrm_interface.c

Signed-off-by: krazey <admin@krazey.de>