BACKPORT: cgroup: Keep accurate count of tasks in each css_set
[ Upstream commit 73a7242a06ff995d771fbe243e72b516feaa6e3d ]

The reference count in the css_set data structure was used as a
proxy of the number of tasks attached to that css_set. However, that
count is actually not an accurate measure especially with thread mode
support. So a new variable nr_tasks is added to the css_set to keep
track of the actual task count. This new variable is protected by
the css_set_lock. Functions that require the actual task count are
updated to use the new variable.

tj: s/task_count/nr_tasks/ for consistency with cgroup_root->nr_cgrps.
    Refreshed on top of cgroup/for-v4.13 which dropped on
    css_set_populated() -> nr_tasks conversion.

Signed-off-by: Waiman Long <longman@redhat.com>
Signed-off-by: Tejun Heo <tj@kernel.org>

[Mathias Gluszczynski: Apply the task accounting to the monolithic 4.9
cgroup core. Decrement the source count before dropping its reference,
as required by the freezer migration hook.]

Signed-off-by: Mathias Gluszczynski <admin@krazey.de>