BACKPORT: bpf: add helper bpf_perf_prog_read_value
[ Upstream commit 4bebdc7a85aa400c0222b5329861e4ad9252f1e5 ]

This patch adds helper bpf_perf_prog_read_cvalue for perf event based bpf
programs, to read event counter and enabled/running time.
The enabled/running time is accumulated since the perf event open.

The typical use case for perf event based bpf program is to attach itself
to a single event. In such cases, if it is desirable to get scaling factor
between two bpf invocations, users can can save the time values in a map,
and use the value from the map and the current value to calculate
the scaling factor.

[Mathias Gluszczynski: Registered the helper only for perf-event
programs because this tree already separates perf-event and
tracepoint dispatch.]

Signed-off-by: Yonghong Song <yhs@fb.com>
Acked-by: Alexei Starovoitov <ast@fb.com>
Acked-by: Daniel Borkmann <daniel@iogearbox.net>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>