BACKPORT: bpf: Add BPF_SOCK_OPS_STATE_CB
[ Upstream commit d44874910a26f3a8f81edf873a2473363f07f660 ]

Adds support for calling sock_ops BPF program when there is a TCP state
change. Two arguments are used; one for the old state and another for
the new state.

There is a new enum in include/uapi/linux/bpf.h that exports the TCP
states that prepends BPF_ to the current TCP state names. If it is ever
necessary to change the internal TCP state values (other than adding
more to the end), then it will become necessary to convert from the
internal TCP state value to the BPF value before calling the BPF
sock_ops function. There are a set of compile checks added in tcp.c
to detect if the internal and BPF values differ so we can make the
necessary fixes.

New op: BPF_SOCK_OPS_STATE_CB.

[Mathias Gluszczynski: Adapted the state callback to the Exynos9810 4.9
TCP state transition paths.]

Signed-off-by: Lawrence Brakmo <brakmo@fb.com>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>

Signed-off-by: Mathias Gluszczynski <admin@krazey.de>