BACKPORT: bpf: Add mtu checking to FIB forwarding helper
[ Upstream commit 4f74fede40df8dbdb5261d50682491126675aac3 ]

Add check that egress MTU can handle packet to be forwarded. If
the MTU is less than the packet length, return 0 meaning the
packet is expected to continue up the stack for help - eg.,
fragmenting the packet or sending an ICMP.

The XDP path needs to leverage the FIB entry for an MTU on the
route spec or an exception entry for a given destination. The
skb path lets is_skb_forwardable decide if the packet can be
sent.

[Mathias Gluszczynski: Adapted IPv4 and IPv6 MTU discovery to the older
FIB and rt6_info route APIs.]

Signed-off-by: David Ahern <dsahern@gmail.com>
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>

Signed-off-by: Mathias Gluszczynski <admin@krazey.de>