BACKPORT: bpf: Change bpf_fib_lookup to return lookup status
[ Upstream commit 4c79579b44b1876444f4d04de31c1a37098a0350 ]

For ACLs implemented using either FIB rules or FIB entries, the BPF
program needs the FIB lookup status to be able to drop the packet.
Since the bpf_fib_lookup API has not reached a released kernel yet,
change the return code to contain an encoding of the FIB lookup
result and return the nexthop device index in the params struct.

In addition, inform the BPF program of any post FIB lookup reason as
to why the packet needs to go up the stack.

The fib result for unicast routes must have an egress device, so remove
the check that it is non-NULL.

[Mathias Gluszczynski: Adapted detailed lookup results and output fields
to the older IPv4 and IPv6 routes.]

Signed-off-by: David Ahern <dsahern@gmail.com>
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>

Signed-off-by: Mathias Gluszczynski <admin@krazey.de>