BACKPORT: bpf: use per htab salt for bucket hash
[ Upstream commit c0203475765f827e7b2eaf0a87222d0766e2cc4b ]

All BPF hash and LRU maps currently have a known and global seed
we feed into jhash() which is 0. This is suboptimal, thus fix it
by generating a random seed upon hashtab setup time which we can
later on feed into jhash() on lookup, update and deletions.

[Mathias Gluszczynski: Applied the per-map hash salt to the
Exynos9810 4.9 hashtab allocation and lookup paths.]

Fixes: 0f8e4bd8a1fc8 ("bpf: add hashtable type of eBPF maps")
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>
Acked-by: Alexei Starovoitov <ast@kernel.org>
Acked-by: Song Liu <songliubraving@fb.com>
Reviewed-by: Eduardo Valentin <eduval@amazon.com>
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>