BACKPORT: bpf: helper to pop data from messages
[ Upstream commit 7246d8ed4dcce23f7509949a77be15fa9f0e3d28 ]

This adds a BPF SK_MSG program helper so that we can pop data from a
msg. We use this to pop metadata from a previous push data call.

[Mathias Gluszczynski: Add the SK_MSG helper, packet-data invalidation,
and TCP send-verdict accounting to this 4.9 sockmap implementation.
The Android UAPI already carries the helper ID and documentation. Omit
the absent kTLS BPF integration hunk.]

Signed-off-by: John Fastabend <john.fastabend@gmail.com>
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>