BACKPORT: bpf: add helper to check for a valid SYN cookie
[ Upstream commit 399040847084a69f345e0a52fd62f04654e0fce3 ]

Using bpf_skc_lookup_tcp it's possible to ascertain whether a packet
belongs to a known connection. However, there is one corner case: no
sockets are created if SYN cookies are active. This means that the final
ACK in the 3WHS is misclassified.

Using the helper, we can look up the listening socket via
bpf_skc_lookup_tcp and then check whether a packet is a valid SYN
cookie ACK.

[Mathias Gluszczynski: Reused the carried ID and adapted SYN-cookie
validation to the older TCP API.]

Signed-off-by: Lorenz Bauer <lmb@cloudflare.com>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>

Signed-off-by: Mathias Gluszczynski <admin@krazey.de>