BACKPORT: bpf: add bpf_skb_adjust_room flag BPF_F_ADJ_ROOM_FIXED_GSO
[ Upstream commit 2278f6cc151a8bef6ba0b3fe3009d14dc3c51c4a ]

bpf_skb_adjust_room adjusts gso_size of gso packets to account for the
pushed or popped header room.

This is not allowed with UDP, where gso_size delineates datagrams. Add
an option to avoid these updates and allow this call for datagrams.

It can also be used with TCP, when MSS is known to allow headroom,
e.g., through MSS clamping or route MTU.

Changes v1->v2:
  - document flag BPF_F_ADJ_ROOM_FIXED_GSO
  - do not expose BPF_F_ADJ_ROOM_MASK through uapi, as it may change.

[Mathias Gluszczynski: Used the Exynos9810 4.9 GSO type bits directly;
UDP_L4 GSO is unavailable in this base, while fixed-size TCP GSO is
supported.]

Link: https://patchwork.ozlabs.org/patch/1052497/
Signed-off-by: Willem de Bruijn <willemb@google.com>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>