BACKPORT: bpf: Validate signals early in bpf_send_signal
[ Upstream commit e1afb70252a8614e1ef7aec05ff1b84fd324b782 ]

Commit 8b401f9ed244 ("bpf: implement bpf_send_signal() helper")
introduced bpf_send_signal() helper. If the context is nmi,
the sending signal work needs to be deferred to irq_work.
If the signal is invalid, the error will appear in irq_work
and it won't be propagated to user.

This patch did an early check in the helper itself to notify
user invalid signal, as suggested by Daniel.

[Mathias Gluszczynski: Applied early signal validation to the adapted
deferred-delivery path.]

Suggested-by: Daniel Borkmann <daniel@iogearbox.net>
Signed-off-by: Yonghong Song <yhs@fb.com>
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>

Signed-off-by: Mathias Gluszczynski <admin@krazey.de>